requirements.txt
pytube
pandas
flask
ffmpeg-python
git+https://github.com/openai/whisper.git